+++ src/Watch/Watch.ts
@@ -1,5 +1,5 @@
 import { bindObserver, destroyWatchers, watchWithScope } from '../helpers'
-import type { Destructor, Observer, Watcher } from '../types'
+import type { Destructor, Observer, Selector, Watcher } from '../types'
 
 /**
  * Watcher class for reactive state tracking.
@@ -20,6 +20,9 @@ export class Watch implements Observer {
   /** Whether the watcher has been destroyed */
   destroyed = false
 
+  /** Tracks if the computation has run at least once. */
+  updated = false
+
   /** Cleanup functions to run when watcher is destroyed */
   readonly destructors = new Set<Destructor>()
 
@@ -32,15 +35,16 @@ export class Watch implements Observer {
     return this.childrenObservers
   }
 
-  constructor (readonly watcher: Watcher<void>, freeParent?: boolean, freeUpdate?: boolean) {
+  constructor (watcher: Selector<void>, freeParent?: boolean, freeUpdate?: boolean)
+  /** @deprecated `update` argument is deprecated, use `Selector` */
+  constructor (watcher: Watcher<void>, freeParent?: boolean, freeUpdate?: boolean)
+  constructor (readonly watcher: Watcher<void> | Selector<void>, freeParent?: boolean, freeUpdate?: boolean) {
     if (!freeParent) {
       bindObserver(this)
     }
 
     if (!freeUpdate) {
-      watchWithScope(this, () => {
-        watcher(false)
-      })
+      this.update()
     }
   }
 
@@ -53,7 +57,8 @@ export class Watch implements Observer {
   update () {
     if (!this.destroyed) {
       watchWithScope(this, () => {
-        this.watcher(true)
+        this.watcher(this.updated) // TODO: remove `this.updated` in major release
+        this.updated = true
       })
     }
   }